Removing some noise
This diff is a little less intellectually pure, but it removes some weird noise.  Firstly, we are basically assuming here that yahoo finance data is a little more trustworthy than some other sources.  Other sources have a nasty habbit of including some holidays as days with the same close price as the previous trading day.  We locate such points and eliminate them if yahoo says there is no data for that day.

Also, just to more easily locate other problems, when we have a discrepency of a penny in the democratic price and the yahoo price we just assume that the yahoo price is actually correct.